+++ src/main/java/com/PickOne/user/model/domain/Password.java
@@ -1,4 +1,4 @@
-package com.PickOne.user.model.domain.user;
+package com.PickOne.user.model.domain;
 
 import com.PickOne.user.service.PasswordEncoder;
 import lombok.AccessLevel;
