+++ src/main/java/com/PickOne/user/service/UserService.java
@@ -1,6 +1,6 @@
 package com.PickOne.user.service;
 
-import com.PickOne.user.model.domain.user.User;
+import com.PickOne.user.model.domain.User;
 
 import java.util.List;
