+++ src/main/java/com/PickOne/user/service/UserServiceImpl.java
@@ -3,9 +3,9 @@ package com.PickOne.user.service;
 import com.PickOne.user.exception.EmailAlreadyExistsException;
 import com.PickOne.user.exception.PasswordMismatchException;
 import com.PickOne.user.exception.UserNotFoundException;
-import com.PickOne.user.model.domain.user.Email;
-import com.PickOne.user.model.domain.user.Password;
-import com.PickOne.user.model.domain.user.User;
+import com.PickOne.user.model.domain.Email;
+import com.PickOne.user.model.domain.Password;
+import com.PickOne.user.model.domain.User;
 import com.PickOne.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
