+++ src/test/java/com/PickOne/user/model/domain/EmailTest.java
@@ -1,6 +1,5 @@
 package com.PickOne.user.model.domain;
 
-import com.PickOne.user.model.domain.user.Email;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
