+++ src/test/java/com/PickOne/user/repository/UserRepositoryTest.java
@@ -2,9 +2,9 @@ package com.PickOne.user.repository;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.PickOne.user.model.domain.user.Email;
-import com.PickOne.user.model.domain.user.Password;
-import com.PickOne.user.model.domain.user.User;
+import com.PickOne.user.model.domain.Email;
+import com.PickOne.user.model.domain.Password;
+import com.PickOne.user.model.domain.User;
 import com.PickOne.user.model.entity.UserEntity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
